Moved load_ndarray() function to utils.array.
